Added Maven Artifact Registry support via Google ADC
