+++ lib/modules/datasource/maven/util.ts
@@ -13,6 +13,7 @@ import { streamToString } from '../../../util/streams';
 import { parseUrl } from '../../../util/url';
 import { normalizeDate } from '../metadata';
 import type { ReleaseResult } from '../types';
+import {getGoogleAccessToken} from "../util";
 import { MAVEN_REPO } from './common';
 import type {
   HttpResourceCheckResult,
@@ -63,11 +64,12 @@ function isUnsupportedHostError(err: { name: string }): boolean {
 
 export async function downloadHttpProtocol(
   http: Http,
-  pkgUrl: URL | string
+  pkgUrl: URL | string,
+  opts: HttpOptions & HttpRequestOptions<string> = {},
 ): Promise<Partial<HttpResponse>> {
   let raw: HttpResponse;
   try {
-    raw = await http.get(pkgUrl.toString());
+    raw = await http.get(pkgUrl.toString(), opts);
     return raw;
   } catch (err) {
     const failedUrl = pkgUrl.toString();
@@ -137,6 +139,21 @@ export async function downloadS3Protocol(pkgUrl: URL): Promise<string | null> {
     }
   }
   return null;
+}
+
+export async function downloadArtifactRegistryProtocol(http: Http, pkgUrl: URL): Promise<Partial<HttpResponse>> {
+  const opts: HttpOptions = {}
+
+  const accessToken = await getGoogleAccessToken();
+  if (accessToken) {
+    const auth = Buffer.from(
+      `${'oauth2accesstoken'}:${accessToken}`
+    ).toString('base64');
+    opts.headers = {authorization: `Basic ${auth}`};
+  }
+  const url = pkgUrl.toString().replace("artifactregistry:", "https:")
+
+  return downloadHttpProtocol(http, url, opts)
 }
 
 async function checkHttpResource(
@@ -259,6 +276,13 @@ export async function downloadMavenXml(
     case 's3:':
       rawContent = (await downloadS3Protocol(pkgUrl)) ?? undefined;
       break;
+    case 'artifactregistry:':
+      ({
+        authorization,
+        body: rawContent,
+        statusCode,
+      } = await downloadArtifactRegistryProtocol(http, pkgUrl));
+      break
     default:
       logger.debug(`Unsupported Maven protocol url:${pkgUrl.toString()}`);
       return {};
